move types to types file
