+++ app/screens/rpc/util/types.ts
@@ -4,7 +4,7 @@ import BN from 'bn.js'
 import { NetworkTokenWithBalance, TokenWithBalance } from 'store/balance'
 import { Network } from '@avalabs/chains-sdk'
 import { FindToken } from 'contracts/contractParsers/utils/useFindToken'
-import { PeerMeta } from 'services/walletconnectv2/WalletConnectService'
+import { PeerMeta } from 'services/walletconnectv2/types'
 import { TransactionParams } from 'store/walletConnectV2/handlers/eth_sendTransaction/utils'
 
 export interface DisplayValueParserProps {
